Move configuration to a file
