+++ src/ProductConstructionService/ProductConstructionService.Api/Configuration/DatabaseConfiguration.cs
@@ -8,7 +8,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace ProductConstructionService.Api;
+namespace ProductConstructionService.Api.Configuration;
 
 internal static class DatabaseConfiguration
 {
