+++ src/ProductConstructionService/ProductConstructionService.Api/Configuration/PcsConfiguration.cs
