+++ src/ProductConstructionService/ProductConstructionService.Api/Program.cs
@@ -5,14 +5,11 @@ using Azure.Identity;
 using Azure.Storage.Queues;
 using Maestro.Data;
 using Microsoft.DotNet.Kusto;
-using Microsoft.Net.Http.Headers;
-using Microsoft.OpenApi.Models;
-using ProductConstructionService.Api;
+using ProductConstructionService.Api.Configuration;
 using ProductConstructionService.Api.Controllers;
 using ProductConstructionService.Api.Queue;
 using ProductConstructionService.Api.Telemetry;
 using ProductConstructionService.Api.VirtualMonoRepo;
-using Swashbuckle.AspNetCore.Swagger;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -52,51 +49,7 @@ else
 builder.Services.AddKustoClientProvider("Kusto");
 builder.AddServiceDefaults();
 builder.Services.AddControllers().EnableInternalControllers();
-
-builder.Services.AddSwaggerGen(options =>
-{
-    options.AddSecurityDefinition(
-        "Bearer",
-        new OpenApiSecurityScheme
-        {
-            Type = SecuritySchemeType.ApiKey,
-            In = ParameterLocation.Header,
-            Scheme = "bearer",
-            Name = HeaderNames.Authorization
-        });
-
-    options.AddSecurityRequirement(new OpenApiSecurityRequirement
-    {
-        {
-            new OpenApiSecurityScheme
-            {   Reference = new OpenApiReference
-                {
-                    Id = "Bearer",
-                    Type = ReferenceType.SecurityScheme
-                }
-            },
-            []
-        }
-    });
-});
-
-builder.Services.Configure<SwaggerOptions>(options =>
-{
-    options.SerializeAsV2 = false;
-    options.PreSerializeFilters.Add(
-        (doc, req) =>
-        {
-            doc.Servers = new List<OpenApiServer>
-            {
-                new()
-                {
-                    Url = $"{req.Scheme}://{req.Host.Value}/",
-                },
-            };
-
-            req.HttpContext.Response.Headers.AccessControlAllowOrigin = "*";
-        });
-});
+builder.ConfigureSwagger();
 
 var app = builder.Build();
